move helper code
